refactor(profiles): match the campaigns list-and-detail pattern
/profile is now an index table like /campaigns (click a name to open,
per-row delete with confirm dialog), with the edit form moved to
/profile/[id] and /profile/new for creation. Agent panel matches
/profile by prefix so subpages keep their context.
